+++ pep249_dbapi/tests/e2e/put_get/put_get_helper.py
@@ -269,87 +269,3 @@ def create_matching_files(
     matching_files = [f"{base_file_name}_{i}.csv" for i in range(1, count + 1)]
     create_test_files(directory, matching_files)
     return matching_files
-
-
-def get_compression_test_file_path(compression_type: str) -> Path:
-    """
-    Get the path for a test file with the specified compression type.
-
-    Args:
-        compression_type: Compression type (GZIP, BZIP2, BROTLI, ZSTD, DEFLATE, RAW_DEFLATE, NONE, LZMA)
-
-    Returns:
-        Path: Path to the test file
-
-    Raises:
-        ValueError: If compression type is not supported
-    """
-    from tests.utils import shared_test_data_dir
-
-    compression_map = {
-        "GZIP": "test_data.csv.gz",
-        "BZIP2": "test_data.csv.bz2",
-        "BROTLI": "test_data.csv.br",
-        "ZSTD": "test_data.csv.zst",
-        "DEFLATE": "test_data.csv.deflate",
-        "RAW_DEFLATE": "test_data.csv.raw_deflate",
-        "NONE": "test_data.csv",
-        "LZMA": "test_data.csv.xz",
-    }
-
-    filename = compression_map.get(compression_type.upper())
-    if not filename:
-        raise ValueError(f"Unsupported compression type: {compression_type}")
-
-    return shared_test_data_dir() / "compression" / filename
-
-
-def create_stage_and_get_compression_file(
-    cursor, stage_prefix: str, compression_type: str
-):
-    """
-    Create a temporary stage and get the compression test file path.
-
-    Args:
-        cursor: Database cursor to use
-        stage_prefix: Prefix for the temporary stage name
-        compression_type: Compression type (GZIP, BZIP2, BROTLI, ZSTD, DEFLATE, RAW_DEFLATE, NONE, LZMA)
-
-    Returns:
-        tuple: (stage_name, test_file_path)
-    """
-    stage_name = create_temporary_stage(cursor, stage_prefix)
-    test_file_path = get_compression_test_file_path(compression_type)
-    return stage_name, test_file_path
-
-
-def assert_put_compression_result(
-    result,
-    expected_source: str,
-    expected_source_compression: str,
-    expected_target: str,
-    expected_target_compression: str,
-):
-    """
-    Assert that PUT result matches expected compression values.
-
-    Args:
-        result: PUT command result row
-        expected_source: Expected source filename
-        expected_source_compression: Expected source compression type
-        expected_target: Expected target filename
-        expected_target_compression: Expected target compression type
-    """
-    assert (
-        result[0] == expected_source
-    ), f"Source should be '{expected_source}', got '{result[0]}'"
-    assert (
-        result[1] == expected_target
-    ), f"Target should be '{expected_target}', got '{result[1]}'"
-    assert (
-        result[4] == expected_source_compression
-    ), f"Source compression should be '{expected_source_compression}', got '{result[4]}'"
-    assert (
-        result[5] == expected_target_compression
-    ), f"Target compression should be '{expected_target_compression}', got '{result[5]}'"
-    assert result[6] == "UPLOADED", f"Status should be 'UPLOADED', got '{result[6]}'"
